requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.1

langchain>=0.2.14
langchain-community>=0.2.12
langchain-openai>=0.1.22

faiss-cpu>=1.8.0

pypdf>=4.2.0
unstructured>=0.14.6
python-docx>=1.1.2
